refactor(ses): move template rendering and the ObjectMapper into SesTemplateService
